pyproject.toml
[project]
name = "cmdbuttons"
version = "0.1.0"
description = "Opens a window with configurable CLI command buttons and stdout"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyqt5",
    "watchdog",
    "pyyaml",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cmdbuttons"]

[project.scripts]
cmdbuttons = "cmdbuttons:main"
